refactor(tracing): address review feedback on span scoping
